Remove IP Geolocation & other customizations
Remove the ability to look up location by IP address. I prefer the
widget using static location. It was possible to simply turn it off
but I also wanted to reduce the size.

Set my preferred settings (colors, icons, location, etc).

Modified so that widget will work when app is set to use a widget
directory that is my git repository containing this repo as a submodule.